feat: clarify the theme controls by consolidating theming into theme and using context for file naming
