+++ app/navigation/index.ts
@@ -1,3 +1,3 @@
 export * from './Primary'
-export * from './root-navigator'
-export * from './navigation-utilities'
+export * from './Root'
+export * from './utilities'
