+++ app/theme/index.ts
@@ -1,4 +1,6 @@
+import _addThemeContext from './paper/addThemeContext'
 export * from './colors'
 export * from './spacing'
 export * from './typography'
 export * from './timing'
+export const addThemeContext = _addThemeContext
