--- app/theming/index.ts
@@ -1,2 +1,0 @@
-import _addThemeContext from './paper/addThemeContext'
-export const addThemeContext = _addThemeContext
